:trollface: return long

+++ stream-plugin/stream-plugin-mybatis-plus/src/main/java/io/github/vampireachao/stream/plugin/mybatisplus/injector/IMapper.java
@@ -16,6 +16,6 @@ public interface IMapper<T> extends BaseMapper<T> {
      * @param list 数据
      * @return 条数
      */
-    int insertOneSql(@Param("list") Collection<T> list);
+    long insertOneSql(@Param("list") Collection<T> list);
 
 }
